pyproject.toml
[project]
name = "modelops"
version = "0.1.0"
description = "Infrastructure orchestration for ML experimentation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "modelops-contracts",
    "rich>=14.1.0",
    "typer>=0.16.1",
    "pyyaml>=6.0",
    "polars>=0.20.0",
    "pyarrow>=14.0.0",
    "dask[distributed]==2024.10.0",
    "pulumi>=3.100.0",
    "pulumi-kubernetes>=4.5.0",
    "kubernetes>=28.0.0",
    "pulumi-azure-native>=3.7.1",
    "numpy>=2.3.2",
    "pulumi-tls>=5.2.2",
    "pulumi-random>=4.18.3",
    "pandas>=2.3.2",
    "azure-storage-blob>=12.26.0",
    "starsim>=3.0.4",
    "sciris>=3.2.4",
]

[project.optional-dependencies]
# Full suite of ModelOps tools for end users
full = [
    "modelops-bundle @ git+https://github.com/institutefordiseasemodeling/modelops-bundle.git",
    "modelops-calabaria @ git+https://github.com/institutefordiseasemodeling/modelops-calabaria.git",
]
# Azure dependencies are in main for MVP
# Future: aws = ["pulumi-aws>=..."]
# Future: gcp = ["pulumi-gcp>=..."]

[project.scripts]
mops = "modelops.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
modelops-contracts = { git = "https://github.com/institutefordiseasemodeling/modelops-contracts.git" }
modelops-bundle = { git = "https://github.com/institutefordiseasemodeling/modelops-bundle.git" }
modelops-calabaria = { git = "https://github.com/institutefordiseasemodeling/modelops-calabaria.git" }

[dependency-groups]
azure = [
    "azure-storage-blob>=12.26.0",
]
bundle = [
    "modelops-bundle",
]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-timeout>=2.4.0",
    "ruff>=0.14.3",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
packages = ["src/modelops", "examples"]

[tool.hatch.build.targets.wheel]
packages = ["src/modelops"]
include = [
    "src/modelops/**/*.py",
    "src/modelops/**/*.yaml",
    "src/modelops/**/*.yml",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: Integration tests that require Dask cluster",
    "slow: Tests that take a long time to run",
]
# Run unit tests by default, integration tests with: pytest -m integration
addopts = "-m 'not integration'"

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "UP",  # pyupgrade (modernize Python code)
    "B",   # flake8-bugbear (detect bugs)
    "SIM", # flake8-simplify (simplify code)
]
ignore = [
    "E722",  # bare-except (TODO: fix gradually)
    "B008",  # function-call-in-default-argument (common in FastAPI/Typer)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["F401", "F811"]  # Allow unused imports and redefinitions in tests
"examples/**/*.py" = ["E402"]  # Allow imports not at top in examples

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
